move components in components folder
