+++ apps/web/app/(app)/onboarding/components/Greeting.tsx
