+++ apps/web/app/(app)/onboarding/components/Objective.tsx
