+++ apps/web/app/(app)/onboarding/components/Onboarding.tsx
