+++ apps/web/app/(app)/onboarding/components/Product.tsx
