+++ apps/web/app/(app)/onboarding/components/Role.tsx
@@ -1,6 +1,6 @@
 "use client";
 
-import { cn } from "@/../../packages/lib/cn";
+import { cn } from "@formbricks/lib/cn";
 import { updateProfileAction } from "@/app/(app)/onboarding/actions";
 import { env } from "@/env.mjs";
 import { createResponse, formbricksEnabled } from "@/lib/formbricks";
